fix(lualine): disable default separators for custom separator (#1233)
I know the commit message might be a bit difficult to grasp, but
I'm struggling to find a clearer way to describe it :( Essentially,
this commit ensures that when our custom separator is used, default
separators (as defined by `section_separators` and `component_separators`)
are not added to the statusline. This prevents certain extreme
scenarios where the default separator cannot be concealed, such as
when in terminal mode with `:ToggleTerm direction=horizontal`.

Signed-off-by: Jint-lzxy <50296129+Jint-lzxy@users.noreply.github.com>
(cherry picked from commit 8268d2768222c9ff37114cb59524cfa44a3f2f7a)