feat(catppuccin): remove all styles of `@property` (#1240)
Based on the upstream spec, `@property` is typically associated with the "key"
in key/value pairs. So IMO since we don't give special treatment to the "values"
(such as assigning them an "italic" property), we should treat the "keys" in the
same manner as well.

Signed-off-by: Jint-lzxy <50296129+Jint-lzxy@users.noreply.github.com>
(cherry picked from commit 092b5a0bafece2409dda15418a2862f2a1180076)